file renaming and re-org
