+++ packages/manager/src/mocks/presets/crud/seeds/index.ts
@@ -1,0 +1,5 @@
+import { linodesSeeder } from './linodes';
+import { placementGroupSeeder } from './placementGroups';
+import { volumesSeeder } from './volumes';
+
+export const dbSeeders = [linodesSeeder, placementGroupSeeder, volumesSeeder];
