+++ packages/manager/src/mocks/presets/extra/regions/legacyRegions.ts
