--- packages/manager/src/mocks/seeds/index.ts
@@ -1,5 +1,0 @@
-import { linodesSeeder } from './linodes-seeder';
-import { placementGroupSeeder } from './placementGroup-seeder';
-import { volumesSeeder } from './volumes-seeder';
-
-export const dbSeeders = [linodesSeeder, placementGroupSeeder, volumesSeeder];
